tryPollFinish
