+++ modules/runtime/src/io/build.io.cpp
@@ -1,5 +1,6 @@
 #include "io.cpp"
 
-#include "ram_io.cpp"
-#include "vfs_ram_reader.cpp"
+#include "ram/ram_io.cpp"
+#include "ram/ram_resolvers.cpp"
+#include "ram/vfs_ram_reader.cpp"
 // #include "dstorage_ram_reader.cpp"
